requirements.txt
fastapi 
sqlalchemy 
psycopg2-binary
uvicorn 
gunicorn 
passlib[bcrypt] 
python-jose 
python-dotenv
dotenv
pydantic
pydantic[email]
pyJWT